[PURPLE-210] 수동 별점 집계용 API 추가
